Stateful feedback page
Moved some logic from the feedbackform to the feedback page so that it (error handling) can be shared between the feedbackform and the select city form on the feedback page.
