+++ takeoutthetrash-ui/src/js/Components/FeedbackForm.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { connect } from "react-redux";
 import * as feedbackActions from "../Actions/Feedback";
 import PropTypes from "prop-types";
@@ -7,8 +7,6 @@ import * as feedbackSelectors from "../Selectors/Feedback";
 import { TextInput, TEXT_AREA } from "../Components/Common/TextInput";
 import * as citiesSelectors from "../Selectors/Cities";
 import { isNilOrEmpty } from "../Utilities/RamdaUtilities";
-
-const COMMENT_MAX = 150;
 
 export const preventFormSubmission = (formValues, city, errors) => {
   return isNilOrEmpty(formValues.comment) ||
@@ -26,7 +24,7 @@ const EmailFormField = ({ visible, formValues, handleInputChange }) => {
     <TextInput
       name="email"
       value={formValues.email}
-      onChange={(e) => handleInputChange(e.target)}
+      onChange={handleInputChange}
       required={false}
       placeholder="Email"
     />
@@ -34,50 +32,23 @@ const EmailFormField = ({ visible, formValues, handleInputChange }) => {
 };
 
 const FeedbackForm = ({
-  postFeedbackForm,
-  feedbackFormValuesUpdated,
   toggledEnableEmailFormField,
   emailFormFieldEnabled,
   city,
+  onChange,
+  formValues,
+  onSubmit,
+  errors,
 }) => {
-  let [formValues, setFormValues] = useState({ comment: "", email: "" });
-  let [errors, setErrors] = useState({});
-
-  const formIsValid = () => {
-    const _errors = {};
-
-    if (formValues.comment.length >= COMMENT_MAX)
-      _errors.comment = `Maximum ${COMMENT_MAX} characters`;
-    //regex validation for email address
-    setErrors(_errors);
-
-    return Object.keys(_errors).length === 0;
-  };
-
-  const handleInputChange = (target) => {
-    const updatedFormValues = {
-      ...formValues,
-      [target.name]: target.value,
-    };
-    setFormValues(updatedFormValues);
-    feedbackFormValuesUpdated(updatedFormValues);
-    if (!formIsValid()) return;
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    postFeedbackForm();
-  };
-
-  const disableSubmitButton = preventFormSubmission(formValues, city, errors);
+  //const disableSubmitButton = preventFormSubmission(formValues, city, errors);
 
   return (
     <div>
-      <form onSubmit={handleSubmit}>
+      <form onSubmit={onSubmit}>
         <TextInput
           name={"comment"}
           value={formValues.comment}
-          onChange={(e) => handleInputChange(e.target)}
+          onChange={onChange}
           required={true}
           placeholder="Comment"
           error={errors.comment}
@@ -95,13 +66,13 @@ const FeedbackForm = ({
         <EmailFormField
           visible={emailFormFieldEnabled}
           formValues={formValues}
-          handleInputChange={handleInputChange}
+          handleInputChange={onChange}
         />
         <div className="feedback-submit-button">
           <button
             type="submit"
             className="submit-button"
-            disabled={disableSubmitButton}
+            //disabled={disableSubmitButton}
           >
             Submit
           </button>
@@ -112,14 +83,22 @@ const FeedbackForm = ({
 };
 
 FeedbackForm.propTypes = {
-  feedbackFormValuesUpdated: PropTypes.func.isRequired,
-  postFeedbackForm: PropTypes.func.isRequired,
   toggledEnableEmailFormField: PropTypes.func.isRequired,
   emailFormFieldEnabled: PropTypes.bool,
   city: PropTypes.shape({
     id: PropTypes.number.isRequired,
     name: PropTypes.string.isRequired,
     rules: PropTypes.arrayOf(PropTypes.object),
+  }),
+  onChange: PropTypes.func.isRequired,
+  onSubmit: PropTypes.func.isRequired,
+  formValues: PropTypes.shape({
+    comment: PropTypes.string,
+    email: PropTypes.string,
+  }),
+  errors: PropTypes.shape({
+    comment: PropTypes.string,
+    email: PropTypes.string,
   }),
 };
 
@@ -129,8 +108,6 @@ const mapStateToProps = (state) => ({
 });
 
 const mapDispatchToProps = {
-  feedbackFormValuesUpdated: feedbackActions.feedbackFormValuesUpdated,
-  postFeedbackForm: feedbackActions.postFeedbackForm,
   toggledEnableEmailFormField: feedbackActions.enableEmailFormFieldToggled,
 };
